Simplified http error translation + renaming
